refactor(hero): relocate state/handlers + resolve typing
